requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
authlib==1.6.5
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==46.0.3
docstring-parser==0.17.0
exceptiongroup==1.3.0 ; python_full_version < '3.11'
fastapi==0.119.1
google-adk==1.6.1
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.122.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-speech==2.34.0
google-cloud-storage==2.19.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.46.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
graphviz==0.21
greenlet==3.2.4 
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.18.0
numpy==2.2.6 ; python_full_version < '3.11'
numpy==2.3.4 ; python_full_version >= '3.11'
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-trace==1.10.0
opentelemetry-resourcedetector-gcp==1.10.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
proto-plus==1.26.1
protobuf==6.33
pydantic==2.12.3
pycryptodome==3.23.0